refactor: cluster vo moved to actor
