+++ src/services/cluster/mod.rs
@@ -1,26 +1,8 @@
 pub mod actor;
-use actor::ClusterActor;
+use actor::{ClusterActor, ClusterCommand};
 use tokio::sync::mpsc::Sender;
 
 use super::stream_manager::interface::TStream;
-
-#[derive(Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub struct PeerAddr(pub String);
-pub enum Connected<T: TStream> {
-    Replica {
-        peer_stream: T,
-        replication_stream: T,
-    },
-    ClusterMemeber {
-        peer_stream: T,
-    },
-}
-
-pub enum ClusterCommand<T: TStream> {
-    AddPeer(PeerAddr, Connected<T>),
-    RemovePeer(PeerAddr),
-    GetPeer(PeerAddr),
-}
 
 pub struct ClusterManager<T: TStream>(Sender<ClusterCommand<T>>);
